+++ packages/screenshoter/src/scroller.js
@@ -8,7 +8,7 @@ function makeScroller({logger, element, scrollingMode = 'mixed'}) {
     moveTo,
     getInnerOffset,
     getSize,
-    getClientRect,
+    getClientRegion,
     getScrollOffset,
     getTranslateOffset,
     getShiftOffset,
@@ -36,8 +36,8 @@ function makeScroller({logger, element, scrollingMode = 'mixed'}) {
     return size
   }
 
-  async function getClientRect() {
-    const region = await element.getClientRect()
+  async function getClientRegion() {
+    const region = await element.getClientRegion()
     // const location = await element.context.getLocationInPage()
     return region
   }
@@ -78,9 +78,12 @@ function makeScroller({logger, element, scrollingMode = 'mixed'}) {
 
   async function scrollTo(offset, element = defaultElement) {
     try {
+      console.log('scrollTo')
       const scrollOffset = await element.scrollTo(offset)
+      console.log(scrollOffset)
       return scrollOffset
     } catch (err) {
+      console.log(err)
       // Sometimes it is expected e.g. on Appium, otherwise, take care
       logger.verbose(`Failed to set current scroll offset!.`, err)
       return {x: 0, y: 0}